renumber compress/decompress
and move around some functions

also tweak code to allow buffer compress/decompress to use same buffer ID for source and destination
